se realizo en trabajo para obtener la lista de personal academico back y fron

+++ gest-ahoex-backend/database/migrations/2024_09_19_225715_create_personal_academico_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->integer('telefono')->unique();
 
-            $table->unsignedInteger('tipo_personal_id');
+            $table->unsignedBigInteger('tipo_personal_id');
 
             $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP'));
